Added type check and data transformation
Type check:
As part of the data cleaning process, some of the values are checked if they are "in the range". For example, month is Jan-Dec, while country code is specified in ISO format. For more information, look at "checkData" function.

Data transformation:
To make sure data is in the correct type (as int, bool, etc...), function (transformDataTypes) now transforms this to correct data type.

Note: new CSV has been added, as "ISO_Code" that is being used to check ISO codes from original data.